Add Guid column and make Raw column nullable
A new Guid column has been added to the WireMockServerMapping model and used in the ServerOrchestrator. The Raw column in the model has been altered to become nullable, resulting in changes in migration scripts and model snapshot. The .gitignore file has also been updated to ignore Sqlite database transaction logs.