fixing #10 - Exception on calling mapping for a not running service (#17)
* Add service orchestration for Mappings

Integrated a ServiceOrchestrator to manage service run checks in Mappings. This ensures that the service is running before certain actions and updates various tests to accommodate the orchestrator. Removed some redundant using directives for cleaner code.

* Fix missing braces in OnGet method

Added missing curly braces in the conditional check within the OnGet method to ensure proper execution flow and code readability. This change prevents potential logical errors and improves the maintainability of the code.

* Add Codacy badge and extend MappingsTest

Added a Codacy badge to the README.md for code quality tracking. Enhanced MappingsTest to include a new test method for the scenario when no service is running. Also refactored ServiceOrchestrator to support mocking in test cases.

* Make `IsRunning` virtual and add unit test for verification

Modified the `IsRunning` property to be virtual in `WireMockService` to allow for testing and extensibility. Added a new unit test in `ServerOrchestratorTest.cs` to verify `IsRunning` behavior with mock configurations.

* Refactor variable declarations to be constants

Changed variable declarations to use 'const' for clarity and immutability in `MappingsTest` and `ServerOrchestratorTest`. This improves code readability and ensures that these values are not altered during execution.