Change reset action to return NoContent and handle UI updates
Updated the reset mapping action to return a NoContentResult instead of redirecting to a page. The Mappings.cshtml now includes a JavaScript function to gray out and disable buttons of a row upon form submission, enhancing user feedback during the reset action. Additionally, formatted JSON output in the WireMockServerMapping class for better readability.