Enhance WireMock with SignalR integration and event system updates (#30)
* Integrate SignalR for real-time mapping updates

Added SignalR to enable real-time communication for mapping updates. Updated Program, ServiceOrchestrator, and Mappings to support SignalR interactions. Included SignalR client library in the Razor page for live updates.

* Add build and CodeQL badges to README

Included badges for build and test status as well as CodeQL scanning to provide better visibility of the project's build and code quality status. This enhances transparency and allows for quick assessments of the project's current health.

* Change reset action to return NoContent and handle UI updates

Updated the reset mapping action to return a NoContentResult instead of redirecting to a page. The Mappings.cshtml now includes a JavaScript function to gray out and disable buttons of a row upon form submission, enhancing user feedback during the reset action. Additionally, formatted JSON output in the WireMockServerMapping class for better readability.

* Update test assertion to check for NoContentResult

Adjusted the `MappingsTest` to assert that `actionResult` is of type `NoContentResult` instead of `RedirectToPageResult`. This change ensures the test accurately reflects the expected behavior of the method under test.

* fixed PR comments

* disable sonar cloud ci cd

* integrated sonarcloud again with pr parameters

* sonar ci cd fixes

* created if than in ci cd of sonar

* added an own PR sonarcloud yaml

* fix naming of sonar on pr yaml

* Add `Html.PathError` to excluded inspections

This change includes `Html.PathError` to the list of excluded inspections within the `qodana.yaml` file. This helps to avoid irrelevant error reporting for specific paths related to WireMock.

* Update GitHub Actions to use setup-dotnet v4

Upgraded actions/setup-dotnet from v1 to v4 in SonarCloud_PR.yml and SonarCloud_push.yml workflows. This ensures compatibility with .NET 8.0.x and improves action reliability and performance.

* Refactor mapping methods and add unit tests

Refactored access modifiers and removed unnecessary attributes in `WireMockServiceList` and `ServiceOrchestrator`. Added unit tests in `ServerOrchestratorTest` to validate the raising of mapping change events.

* Increase coverage target to 80% and simplify event handler.

Raised the coverage target from 50% to 80% in the `codecov.yml` configuration file. Additionally, simplified the event handler syntax in `ServerOrchestratorTest.cs` and removed an extraneous newline character.